GLUI: use double time in seconds for pugl events

+++ glui/smwindow.hh
@@ -39,7 +39,7 @@ protected:
   bool                      update_full_redraw = false;
   bool                      debug_update_region = false;
   EventLoop                *m_event_loop = nullptr;
-  unsigned long             last_click_time = 0;
+  double                    last_click_time = 0;
   unsigned                  last_click_button = 0;
   Point                     last_click_pos;
   unsigned                  mouse_buttons_pressed = 0;
